requirements.txt
altgraph==0.17.5
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
asgiref==3.11.1
boto3==1.42.54
botocore==1.42.54
bottle==0.13.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
clr_loader==0.2.10
colorama==0.4.6
distro==1.9.0
Django==6.0.2
django-storages==1.14.6
docstring_parser==0.17.0
et_xmlfile==2.0.0
git-filter-repo==2.47.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jmespath==1.1.0
lxml==6.0.2
openai==2.21.0
openpyxl==3.1.5
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
proxy_tools==0.1.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
pypdf==6.7.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pythonnet==3.0.5
pywebview==6.1
pywin32-ctypes==0.2.3
requests==2.32.5
s3transfer==0.16.0
setuptools==82.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.11.0
